+++ src/types/lyric.ts
@@ -5,6 +5,8 @@ export type Lyric = {
     id?: string;
     provider?: Provider;
     exception?: Exception;
-    syncedLyric?: string;
-    plainLyric?: string;
+    syncedLyric?: string[];
+    plainLyric?: string[];
+    lang?: string;
+    instrumental?: boolean;
 }